some funcs are moved to utils.py
