+++ unihub/utils.py
@@ -1,5 +1,8 @@
 from PIL import Image
 from io import BytesIO
+from rest_framework.authtoken.models import Token
+from rest_framework.response import Response
+from rest_framework import status
 
 def validate_png(file_bytes):
     try:
@@ -9,3 +12,30 @@ def validate_png(file_bytes):
         return True, None
     except Exception as e:
         return False, f"Error processing image: {str(e)}"
+    
+def serializers_get_user_from_request(self):
+    request = self.context.get('request')
+
+    if not request:
+        return None
+    
+    token = request.COOKIES.get('token')
+
+    if not token:
+        return None
+
+    try:
+        return Token.objects.get(key=token).user
+    except Token.DoesNotExist:
+        return None
+    
+def get_user_from_request(request):
+    token = request.COOKIES.get("token")
+
+    if not token:
+        return None, Response({"error": "No token provided"}, status=status.HTTP_401_UNAUTHORIZED)
+    try:
+        user = Token.objects.get(key=token).user
+        return user, None
+    except Token.DoesNotExist:
+        return None, Response({"error": "Invalid token"}, status=status.HTTP_401_UNAUTHORIZED)
